+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog for NeoFS Node
 - EC parts of expired object no longer remain on SN (#3718)
 - Metabase reset not deleting old buckets in some cases (#3744)
 - Object left in the storage after unsuccessful meta PUT (#3744)
+- Object left in some shards on fatal broadcast failure (#3744)
 
 ### Changed
 - Optimized locking info in metabase (#3672)